Moved hystrix/servo/statsd/graphite config from util to new module: hystrix
