+++ hystrix/src/main/java/com/redhat/lightblue/hystrix/ServoGraphiteSetup.java
@@ -16,14 +16,14 @@
  You should have received a copy of the GNU General Public License
  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.redhat.lightblue.util;
+package com.redhat.lightblue.hystrix;
 
 import com.netflix.hystrix.Hystrix;
 import com.netflix.hystrix.contrib.servopublisher.HystrixServoMetricsPublisher;
 import com.netflix.hystrix.strategy.HystrixPlugins;
 import com.netflix.servo.publish.*;
 import com.netflix.servo.publish.graphite.GraphiteMetricObserver;
-import com.redhat.lightblue.util.statsd.StatsdMetricObserver;
+import com.redhat.lightblue.hystrix.statsd.StatsdMetricObserver;
 import java.util.ArrayList;
 import java.util.List;
 import org.slf4j.Logger;
