+++ hystrix/src/main/java/com/redhat/lightblue/hystrix/statsd/StatsdMetricObserver.java
@@ -16,7 +16,7 @@
  You should have received a copy of the GNU General Public License
  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.redhat.lightblue.util.statsd;
+package com.redhat.lightblue.hystrix.statsd;
 
 import java.util.List;
