+++ hystrix/src/test/java/com/redhat/lightblue/hystrix/ServoGraphiteSetupTest.java
@@ -16,11 +16,11 @@
  You should have received a copy of the GNU General Public License
  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.redhat.lightblue.util;
+package com.redhat.lightblue.hystrix;
 
 import com.netflix.servo.publish.MetricObserver;
 import com.netflix.servo.publish.graphite.GraphiteMetricObserver;
-import com.redhat.lightblue.util.statsd.StatsdMetricObserver;
+import com.redhat.lightblue.hystrix.statsd.StatsdMetricObserver;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.Assert;
